requirements.txt
streamlit
tensorflow
Pillow
numpy
huggingface-hub